pyproject.toml
[build-system]
requires = ["setuptools>=42", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# For smarter version schemes and other configuration options,
# check out https://github.com/pypa/setuptools_scm
version_scheme = "post-release"
local_scheme = "no-local-version"

[project]
name = "pymeshup-gui"
dynamic = ["version"]
description = "Make volume meshes the easy way"
authors = [{ name = "Ruben de Bruin", email = "rubendebruin@gmail.com" }]
license = "MPL-2.0"
readme = "README.md"
keywords = ["meshing", "volume", "geometry"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "vedo",
    "pymeshup",
    "PySide6-Essentials",
    "pymeshlab",
    "capytaine>=2.2",
    "mafredo",
    "netcdf4>=1.7.2",
    "pyinstaller"
]
urls = { "Documentation" = "https://open-ocean.org/" }

[project.optional-dependencies]
testing = ["setuptools", "pytest"]
exe = ["pyinstaller"]

[project.scripts]
pymeshup-gui = "pymeshup_gui.gui.main:run"

[dependency-groups]
dev = [
    "pre-commit>=3.5.0",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
]


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = ["src/**", "tests/**", "docs/**", "README.md", "LICENSE.txt"]

[tool.hatch.build.targets.wheel]
include = ["src/**", "tests/**", "docs/**", "README.md", "LICENSE.txt"]

[tool.pytest.ini_options]
addopts = ["-m", "not interactive", "-p", "no:cov"]
markers = [
    "interactive: mark tests as interactive (skipped by default)",
    "gui: mark tests that launch a GUI (skipped by default)",
]
testpaths = ["tests"]
norecursedirs = ["dist", "build", ".tox"]


[tool.ruff]
# Geldt voor zowel lint (check) als format
exclude = ["src/pymeshup_gui/DAVE/GHS_to_DAVE_form.py"]

[tool.ruff.lint.per-file-ignores]
"src/pymeshup_gui/DAVE/GHS_to_DAVE_form.py" = ["ALL"]
